builder_cloud/gateway/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
azure-storage-blob>=12.19.0
python-dotenv>=1.0.0
pydantic>=2.0.0
